+++ packages/discovery-react-components/src/components/StructuredQuery/__stories__/StructuredQuery.stories.tsx
@@ -7,13 +7,25 @@ import StructuredQuery from '../StructuredQuery';
 import { defaultMessages } from '../messages';
 import marked from 'marked';
 import defaultReadme from './default.md';
+import { createDummyResponsePromise } from 'utils/testingUtils';
+import DiscoveryV2 from 'ibm-watson/discovery/v2';
+import { action } from '@storybook/addon-actions';
 
 const props = () => ({
   messages: object("Default messages for the component's text strings", defaultMessages)
 });
 
+class DummySearchClientReturnsFields extends DummySearchClient {
+  public async listFields(fetchFieldsParams: DiscoveryV2.ListFieldsParams): Promise<any> {
+    action('listFields')(fetchFieldsParams);
+    return createDummyResponsePromise({
+      fields: [{ field: 'field name 1' }, { field: 'field name 2' }]
+    });
+  }
+}
+
 const discoverySearchProps = (): DiscoverySearchProps => ({
-  searchClient: new DummySearchClient(),
+  searchClient: new DummySearchClientReturnsFields(),
   projectId: text('Project ID', 'project-id')
 });