+++ examples/discovery-search-app/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1-beta.5](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.4...discovery-search-app@1.0.1-beta.5) (2020-01-28)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.1-beta.4](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.2...discovery-search-app@1.0.1-beta.4) (2020-01-23)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
 ## [1.0.1-beta.3](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.2...discovery-search-app@1.0.1-beta.3) (2020-01-22)
 
 **Note:** Version bump only for package discovery-search-app